+++ R/class.R
@@ -74,25 +74,26 @@ validate_messydate <- function(x) {
 
   if (any(grepl("[A-WYZa-z]", values))) {
     stop(
-      "The only alpha character allowed in messy dates is 'X' for unspecified time components",
+      "The only alpha character allowed in messy dates is 'X' for
+      unspecified time components",
       call. = FALSE
     )
   }
 
   if (!all(grepl("[0-9]", values))) {
     stop(
-      "Messy dates require at least one specified date component.",
+      "Messydt object requires at least one specified date component.",
       call. = FALSE
     )
     }
 
   if (any(grepl("!|\\(|\\)|\\+|\\=|\\/|,|;|>|<|_|\\^|'|&|\\$|#", values))) {
     stop(
-      "Messy dates can only consist of numbers and some special symbols: []{}..X%?~",
+      "Messydt object can only consist of numbers and
+      some special symbols: []{}..X%?~",
       call. = FALSE
     )
   }
-
   x
 }