+++ R/coerce_to_messydate.R
@@ -5,6 +5,7 @@
 #' such as `Date`, `POSIXct`, `POSIXlt`, or character.
 #' @return A `messydt` class object
 #' @examples
+#' as_messydate("28 BC")
 #' as_messydate("2021")
 #' as_messydate("2021-02")
 #' as_messydate("2021-02-01")
@@ -40,22 +41,21 @@ as_messydate.POSIXlt <- function(x) {
 #' @describeIn as_messydate Coerce character date objects to `messydt` class
 #' @export
 as_messydate.character <- function(x) {
-
-  d <- x
-  d <- standardise_date_separators(d)
+  d <- standardise_date_separators(x)
   d <- standardise_date_order(d)
+  d <- standardise_date_input(d)
   d <- standardise_unspecifieds(d)
   d <- standardise_widths(d)
   d <- standardise_ranges(d)
   d <- remove_imprecision(d)
-
   new_messydate(d)
 }
 
 # Helper functions
 standardise_date_separators <- function(dates) {
   dates <- stringr::str_replace_all(dates,
-                                    "([:digit:]{4})([:digit:]{2})([:digit:]{2})",
+                                    "([:digit:]{4})([:digit:]{2})
+                                    ([:digit:]{2})",
                                     "\\1-\\2-\\3")
   dates <- stringr::str_replace_all(dates,
                                     "(?<=[:digit:])\\.(?=[:digit:])",
@@ -67,15 +67,28 @@ standardise_date_separators <- function(dates) {
 
 standardise_date_order <- function(dates) {
   dates <- stringr::str_replace_all(dates,
-                                    "([:digit:]{2})-([:digit:]{2})-([:digit:]{4})",
-                                    "\\3-\\2-\\1")
+                                    "([:digit:]{2})-([:digit:]{2})-
+                                    ([:digit:]{4})", "\\3-\\2-\\1")
   dates
 }
 
-standardise_widths <- function(dates) {
-  dates <- stringr::str_replace_all(dates, "-([:digit:])$", "-0\\1")
-  dates <- stringr::str_replace_all(dates, "-([:digit:])-", "-0\\1-")
-  dates <- stringr::str_replace_all(dates, "^([:digit:])-", "0\\1-")
+standardise_date_input <- function(dates) {
+  as_bc_dates <- function(dates) {
+    # remove before christ letters
+    dates <- stringr::str_remove_all(dates, "(bc|BC|Bc|bC)")
+    dates <- stringr::str_trim(dates, side = "both")
+    dates <- paste0("-", dates)
+    dates
+  }
+  as_ac_dates <- function(dates) {
+    # remove after christ letters
+    dates <- stringr::str_remove_all(dates, "(ad|AD|Ad|aD)")
+    dates <- stringr::str_trim(dates, side = "both")
+  }
+  dates <- ifelse(stringr::str_detect(dates, "(bc|BC|Bc|bC)"),
+                  as_bc_dates(dates), dates)
+  dates <- ifelse(stringr::str_detect(dates, "(ad|AD|Ad|aD)"),
+                  as_ac_dates(dates), dates)
   dates
 }
 
@@ -83,9 +96,38 @@ standardise_unspecifieds <- function(dates) {
   dates <- stringr::str_replace_all(dates, "^NA", "XXXX")
   dates <- stringr::str_replace_all(dates, "-NA", "-XX")
   dates <- stringr::str_replace_all(dates, "0000", "XXXX")
-  dates <- stringr::str_replace_all(dates, "-00", "-XX")
+  dates <- stringr::str_replace_all(dates, "-00-|-0-|-0$|-00$", "-XX-")
   dates <- stringr::str_replace_all(dates, "\\?\\?\\?\\?", "XXXX")
   dates <- stringr::str_replace_all(dates, "-\\?\\?", "-XX")
+  dates
+}
+
+standardise_widths <- function(dates) {
+  # Negative year only
+  dates <- stringr::str_replace_all(dates, "^-([:digit:]{1})$", "-000\\1")
+  dates <- stringr::str_replace_all(dates, "^-([:digit:]{2})$", "-00\\1")
+  dates <- stringr::str_replace_all(dates, "^-([:digit:]{3})$", "-0\\1")
+  # Uncertain and aproximate year only
+  dates <- stringr::str_replace_all(dates, "^~([:digit:]{1})$", "000\\1~")
+  dates <- stringr::str_replace_all(dates, "^~([:digit:]{2})$", "00\\1~")
+  dates <- stringr::str_replace_all(dates, "^~([:digit:]{3})$", "0\\1~")
+  dates <- stringr::str_replace_all(dates, "^\\?([:digit:]{1})$", "000\\1?")
+  dates <- stringr::str_replace_all(dates, "^\\?([:digit:]{2})$", "00\\1?")
+  dates <- stringr::str_replace_all(dates, "^\\?([:digit:]{3})$", "0\\1?")
+  dates <- ifelse(stringr::str_detect(dates, "^([:digit:]{1})~$|^([:digit:]{1})\\?$"),
+                  paste0("000", dates), dates)
+  dates <- ifelse(stringr::str_detect(dates, "^([:digit:]{2})~$|^([:digit:]{2})\\?$"),
+                  paste0("00", dates), dates)
+  dates <- ifelse(stringr::str_detect(dates, "^([:digit:]{3})~$|^([:digit:]{3})\\?$"),
+                  paste0("0", dates), dates)
+  # Year only
+  dates <- stringr::str_replace_all(dates, "^([:digit:]{1})$", "000\\1")
+  dates <- stringr::str_replace_all(dates, "^([:digit:]{2})$", "00\\1")
+  dates <- stringr::str_replace_all(dates, "^([:digit:]{3})$", "0\\1")
+  # Adds zero padding to months and days
+  dates <- stringr::str_replace_all(dates, "-([:digit:])$", "-0\\1")
+  dates <- stringr::str_replace_all(dates, "-([:digit:])-", "-0\\1-")
+  dates <- stringr::str_replace_all(dates, "^([:digit:])-", "0\\1-")
   dates
 }
 
@@ -97,23 +139,5 @@ standardise_ranges <- function(dates) {
 
 remove_imprecision <- function(dates) {
   dates <- stringr::str_replace_all(dates, "-XX$", "")
-  dates <- stringr::str_replace_all(dates, "-XX$", "")
   dates
-}
-
-standardise_date_input <- function(dates) {
-
-  as_bc_dates <- function(dates) {
-    dates <- stringr::str_remove_all(dates, "(bc|BC|Bc|bC)")
-    # remove before christ letters
-    dates <- paste0("-", dates) # adds a negative sign to date
-    dates
-  }
-  dates <- stringr::str_remove_all(dates, "(ad|AD|Ad|aD)")
-  # remove after christ
-  dates <- ifelse(stringr::str_detect(dates, "(bc|BC|Bc|bC)"),
-                  as_bc_dates(dates), dates)
-  # replacing BC for corresponding negative dates
-  dates <- stringr::str_trim(dates, side = "both")
-  # removes trailing white spaces
 }