+++ R/messydate_make.R
@@ -1,7 +1,7 @@
 #' Make messy dates from multiple variables
 #'
 #' Transforms multiple date inputs contained in different columns into one.
-#' @param ... One (ymd) or three (yyyy, mm, dd) variables
+#' @param ... One (yyyy-mm-dd) or three (yyyy, mm, dd) variables
 #' @importFrom purrr map
 #' @return A character vector containing the
 #' @examples
@@ -15,6 +15,8 @@ make_messydate <- function(...) {
   } else if (length(dots) == 3) {
     dots <- purrr::map(dots, as.character)
     dates <- unlist(purrr::pmap_chr(dots, paste, sep = "-"))
-  } else stop("Either you need to pass make_messydate() one variable (i.e. 'yyyy-mm-dd' three (yyyy, mm, dd).")
-  new_messydate(dates)
+    dates <- gsub("NA-NA-NA", "NA", dates)
+  } else stop("Pass make_messydate() one variable (yyyy-mm-dd)
+              or three variables (yyyy, mm, dd).")
+  as_messydate(dates)
 }