+++ tests/testthat/test-messydate-make.R
@@ -1,7 +1,7 @@
 test_that("mutiple variables are properly bind", {
-  expect_equal(make_messydate("2010", "10", "10"), as_messydate("2010-10-10"))
+  expect_equal(make_messydate("2010", "1", "1"), as_messydate("2010-01-01"))
   expect_equal(make_messydate("2010", "XX", "10"), as_messydate("2010-XX-10"))
-  expect_equal(make_messydate("2010-10-10"), as_messydate("2010-10-10"))
+  expect_equal(make_messydate("2010-01-01"), as_messydate("2010-1-1"))
   expect_equal(make_messydate("2012?", "01", "01"), as_messydate("2012?-01-01"))
   expect_equal(make_messydate("2012", "01", "01~"), as_messydate("2012-01-01~"))
 })